ui updates
Getting tired of documentation and want to go to sleep so if you have
questions send me a message. will need coffee.

For the most part this update includes files i missed in other commits
but referenced and some major changes including:
Seperating the tabbar into its own module
Changing many of the templates related to displaying nicks so they are
atomic and have descriptive classes. Causes templates to be built more
manually but improves control and makes the process cleaner (see
ui.Theme and util.urlifier + ircnick and nickbtn templates)
Nicks in windows now have context menus
Some detached window optimizations (mainly css stuff)
Refactored the nick menu building to a template and executing is no
longer stored with in the element but the command is built in the
template and retrieved from a data item

That might not be all folks but im done for now